# StudentPortal.API/bruno/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StudentPortal API",
    "description": "PRN232 Lab 1 – RESTful API for managing Students, Courses, Subjects and Semesters.",
    "contact": {
      "name": "PRN232 Student",
      "email": "student@fpt.edu.vn"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/subjects": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "summary": "Get all subjects with pagination, search, sorting, select, and expand options.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortdescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$expand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a paged list of subjects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectPagedResponseApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subjects"
        ],
        "summary": "Create a new subject.",
        "requestBody": {
          "description": "Subject data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subject created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or subject code already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/subjects/{id}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "summary": "Get a single subject by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subject found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponseApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subjects"
        ],
        "summary": "Update an existing subject.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subject updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or code conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subjects"
        ],
        "summary": "Delete a subject by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subject deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateSubjectRequest": {
        "required": [
          "credit",
          "subjectCode",
          "subjectName"
        ],
        "type": "object",
        "properties": {
          "subjectCode": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "example": "PRN232"
          },
          "subjectName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "example": "ASP.NET Web API"
          },
          "credit": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 3
          }
        },
        "additionalProperties": false,
        "description": "Payload for creating a new Subject."
      },
...
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "ObjectPagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Pagination metadata wrapper."
      },
      "ObjectPagedResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ObjectPagedResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
      "SubjectResponse": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "subjectCode": {
            "type": "string",
            "nullable": true
          },
          "subjectName": {
            "type": "string",
            "nullable": true
          },
          "credit": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Base subject response — no navigation properties."
      },
      "SubjectResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/SubjectResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
      "UpdateSubjectRequest": {
        "required": [
          "credit",
          "subjectCode",
          "subjectName"
        ],
        "type": "object",
        "properties": {
          "subjectCode": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "example": "PRN232"
          },
          "subjectName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "example": "ASP.NET Web API"
          },
          "credit": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "example": 3
          }
        },
        "additionalProperties": false,
        "description": "Payload for updating an existing Subject."
      }
    }
  }
}